Implemented CNN + NEP classifier from paper.
Got slightly better results than paper for vanilla CNN classifier, and
slightly worse results than paper for NEP classifier. NEP is still and
improvement over vanilla though, it's just that the gain isn't as
dramatic as in the paper. May be able to play around with this a bit
more to get more improvement, but for now I think I'm going to leave it
and try other things.